ubits

Module examples

Examples of generated output using different configurations.

This module is only compiled with documentation builds and will not be available for import.

use the bitfield macro to generate a field struct and flag enum.

Modules

flagless_example

Example of generated output, using only a bit field without a flag enum.

full_example

Example of generated output, using both a bit field and a flag enum.