Module ubits::examples::full_example

Example of generated output, using both a bit field and a flag enum.

This module is only compiled with documentation builds and will not be available for import.

use the bitfield macro to generate a field struct and flag enum.

Structs

ExampleField

Optional docstring for my example bit field.

Enums

ExampleFlags

Optional docstring for my flag enum.